Closes Issue #11. Put startup inside a timeout.
The coffeescript package has an issue where Template is not defined at
the time of compilation. So, as a workaround, users are putting their
Template code inside a Meteor.startup callback. The problem is that this
user code is getting executed AFTER our page machinery starts. So it
causes some weird behavior when you navigate directly to a page. To fix
it, I put the startup call inside a 0ms timeout to give user code a
chance to run first. Little hacky. Should probably submit an issue to
the Meteor team.