adding the Meteor.currentPage() convenience method
Now instead of having to do Meteor.router.invocation() you can just do Meteor.currentPage()

This small change makes it simpler and supported to access and manipulate the current page's dictionary. It's my understanding that this dictionary is the presentation model for the current page and so it makes sense to keep presentation only variables (like currentlySelected="asdf32asd" or editing=true) in this hash, but Meteor.router.invocation() all over the code seems brittle.