reactive before filter redirects are now torn down completely
This fixes a bug where a page invocation is re-run as a result of reactivity after it should have been torn down due to a redirect. It prevents unexpected behavior and extraneous redraws.

The issue was that the Deps.Computation instance that was supposed to be stopped when the redirect occurred was not available outside of its scope yet. However, Deps.autorun makes the computation available inside the scope of the autorun function as a parameter and so can be stopped that way if necessary, which is the real change here.